+++ libs/fairness/src/lib/FairnessWizard.tsx
@@ -16,13 +16,11 @@ import React from "react";
 
 import { EmptyHeader } from "./Controls/EmptyHeader";
 import { FairnessWizardStyles } from "./FairnessWizard.styles";
-import { getFairnessWizardState } from "./FairnessWizardState";
+import { getFairnessWizardState, IWizardStateV2 } from "./FairnessWizardState";
 import { FairnessWizardTabs } from "./FairnessWizardTabs";
 import { IFairnessProps } from "./IFairnessProps";
 import { IFairnessOption } from "./util/FairnessMetrics";
 import { IBinnedResponse } from "./util/IBinnedResponse";
-import { IFairnessContext } from "./util/IFairnessContext";
-import { MetricsCache } from "./util/MetricsCache";
 import { IPerformanceOption } from "./util/PerformanceMetrics";
 import { WizardBuilder } from "./util/WizardBuilder";
 
@@ -48,27 +46,6 @@ export interface IFeatureBinPickerPropsV2 {
   selectedBinIndex: number;
   onBinChange: (index: number) => void;
 }
-
-export interface IWizardStateV2 {
-  showIntro: boolean;
-  activeTabKey: string;
-  selectedModelId?: number;
-  dashboardContext: IFairnessContext;
-  performanceMetrics: IPerformanceOption[];
-  fairnessMetrics: IFairnessOption[];
-  selectedPerformanceKey: string | undefined;
-  selectedFairnessKey: string | undefined;
-  errorBarsEnabled: boolean;
-  featureBins: IBinnedResponse[];
-  selectedBinIndex: number;
-  metricCache: MetricsCache;
-}
-
-export const introTabKey = "introTab";
-export const featureBinTabKey = "featureBinTab";
-export const performanceTabKey = "performanceTab";
-export const fairnessTabKey = "fairnessTab";
-export const reportTabKey = "reportTab";
 
 export const flights = {
   skipFairness: false
